+++ src/useCss.ts
@@ -2,7 +2,8 @@ import { create, NanoRenderer } from 'nano-css';
 import { addon as addonCSSOM, CSSOMAddon } from 'nano-css/addon/cssom';
 import { addon as addonVCSSOM, VCSSOMAddon } from 'nano-css/addon/vcssom';
 import { cssToTree } from 'nano-css/addon/vcssom/cssToTree';
-import { useLayoutEffect, useMemo } from 'react';
+import { useMemo } from 'react';
+import useIsomorphicLayoutEffect from './useIsomorphicLayoutEffect';
 
 type Nano = NanoRenderer & CSSOMAddon & VCSSOMAddon;
 const nano = create() as Nano;
@@ -15,7 +16,7 @@ const useCss = (css: object): string => {
   const className = useMemo(() => 'react-use-css-' + (counter++).toString(36), []);
   const sheet = useMemo(() => new nano.VSheet(), []);
 
-  useLayoutEffect(() => {
+  useIsomorphicLayoutEffect(() => {
     const tree = {};
     cssToTree(tree, css, '.' + className, '');
     sheet.diff(tree);