+++ stories/createGlobalState.story.tsx
@@ -3,27 +3,43 @@ import React, { FC } from "react";
 import { createGlobalState } from "../src";
 import ShowDocs from "./util/ShowDocs";
 
-const useGlobalValue = createGlobalState<number>(0);
+const useMyGlobalState = createGlobalState<number>(0);
 
-const CompA: FC = () => {
-  const [value, setValue] = useGlobalValue();
+const SetValueDirectly: FC = () => {
+  const [value, setValue] = useMyGlobalState(2);
 
-  return <button onClick={() => setValue(value + 1)}>+</button>;
-};
-
-const CompB: FC = () => {
-  const [value, setValue] = useGlobalValue();
-
-  return <button onClick={() => setValue(value - 1)}>-</button>;
-};
-
-const Demo: FC = () => {
-  const [value] = useGlobalValue();
   return (
     <div>
       <p>{value}</p>
-      <CompA />
-      <CompB />
+      <div>
+        <button onClick={() => setValue(value + 1)}>+</button>
+        <button onClick={() => setValue(value - 1)}>-</button>
+      </div>
+    </div>
+  );
+};
+
+const SetValueWithFunctionalUpdate: FC = () => {
+  const [value, setValue] = useMyGlobalState(3);
+
+  return (
+    <div>
+      <p>{value}</p>
+      <div>
+        <button onClick={() => setValue(val => val + 1)}>+</button>
+        <button onClick={() => setValue(val => val - 1)}>-</button>
+      </div>
+    </div>
+  );
+};
+
+const Demo: FC = () => {
+  const [value] = useMyGlobalState(1);
+  return (
+    <div>
+      <p>{value}</p>
+      <SetValueDirectly />
+      <SetValueWithFunctionalUpdate />
     </div>
   );
 };