+++ tests/setupTests.ts
@@ -1,0 +1,6 @@
+import 'jest-localstorage-mock';
+
+(window as any).ResizeObserver = class ResizeObserver {
+  observe() {}
+  disconnect() {}
+};