+++ tests/useMeasure.test.ts
@@ -1,4 +1,3 @@
-/* eslint-disable */
 import { renderHook, act } from '@testing-library/react-hooks';
 import useMeasure, { UseMeasureRef } from '../src/useMeasure';
 
@@ -11,17 +10,17 @@ it('by default, state defaults every value to -1', () => {
   });
 
   expect(result.current[1]).toMatchObject({
-    width: -1,
-    height: -1,
-    top: -1,
-    bottom: -1,
-    left: -1,
-    right: -1,
+    width: 0,
+    height: 0,
+    top: 0,
+    bottom: 0,
+    left: 0,
+    right: 0,
   });
 });
 
 it('synchronously sets up ResizeObserver listener', () => {
-  let listener: ((rect: any) => void) | undefined;
+  let listener: ((rect: any) => void) | undefined = undefined;
   (window as any).ResizeObserver = class ResizeObserver {
     constructor(ls) {
       listener = ls;
@@ -31,7 +30,7 @@ it('synchronously sets up ResizeObserver listener', () => {
   };
 
   const { result } = renderHook(() => useMeasure());
-
+  
   act(() => {
     const div = document.createElement('div');
     (result.current[0] as UseMeasureRef)(div);
@@ -41,7 +40,7 @@ it('synchronously sets up ResizeObserver listener', () => {
 });
 
 it('tracks rectangle of a DOM element', () => {
-  let listener: ((rect: any) => void) | undefined;
+  let listener: ((rect: any) => void) | undefined = undefined;
   (window as any).ResizeObserver = class ResizeObserver {
     constructor(ls) {
       listener = ls;
@@ -51,27 +50,25 @@ it('tracks rectangle of a DOM element', () => {
   };
 
   const { result } = renderHook(() => useMeasure());
-
+  
   act(() => {
     const div = document.createElement('div');
     (result.current[0] as UseMeasureRef)(div);
   });
-
+  
   act(() => {
-    listener!([
-      {
-        contentRect: {
-          x: 1,
-          y: 2,
-          width: 200,
-          height: 200,
-          top: 100,
-          bottom: 0,
-          left: 100,
-          right: 0,
-        },
-      },
-    ]);
+    listener!([{
+      contentRect: {
+        x: 1,
+        y: 2,
+        width: 200,
+        height: 200,
+        top: 100,
+        bottom: 0,
+        left: 100,
+        right: 0,
+      }
+    }]);
   });
 
   expect(result.current[1]).toMatchObject({
@@ -87,7 +84,7 @@ it('tracks rectangle of a DOM element', () => {
 });
 
 it('tracks multiple updates', () => {
-  let listener: ((rect: any) => void) | undefined;
+  let listener: ((rect: any) => void) | undefined = undefined;
   (window as any).ResizeObserver = class ResizeObserver {
     constructor(ls) {
       listener = ls;
@@ -97,27 +94,25 @@ it('tracks multiple updates', () => {
   };
 
   const { result } = renderHook(() => useMeasure());
-
+  
   act(() => {
     const div = document.createElement('div');
     (result.current[0] as UseMeasureRef)(div);
   });
-
+  
   act(() => {
-    listener!([
-      {
-        contentRect: {
-          x: 1,
-          y: 1,
-          width: 1,
-          height: 1,
-          top: 1,
-          bottom: 1,
-          left: 1,
-          right: 1,
-        },
-      },
-    ]);
+    listener!([{
+      contentRect: {
+        x: 1,
+        y: 1,
+        width: 1,
+        height: 1,
+        top: 1,
+        bottom: 1,
+        left: 1,
+        right: 1,
+      }
+    }]);
   });
 
   expect(result.current[1]).toMatchObject({
@@ -130,22 +125,20 @@ it('tracks multiple updates', () => {
     left: 1,
     right: 1,
   });
-
+  
   act(() => {
-    listener!([
-      {
-        contentRect: {
-          x: 2,
-          y: 2,
-          width: 2,
-          height: 2,
-          top: 2,
-          bottom: 2,
-          left: 2,
-          right: 2,
-        },
-      },
-    ]);
+    listener!([{
+      contentRect: {
+        x: 2,
+        y: 2,
+        width: 2,
+        height: 2,
+        top: 2,
+        bottom: 2,
+        left: 2,
+        right: 2,
+      }
+    }]);
   });
 
   expect(result.current[1]).toMatchObject({
@@ -170,7 +163,7 @@ it('calls .disconnect() on ResizeObserver when component unmounts', () => {
   };
 
   const { result, unmount } = renderHook(() => useMeasure());
-
+  
   act(() => {
     const div = document.createElement('div');
     (result.current[0] as UseMeasureRef)(div);