+++ tests/useSessionStorage.test.ts
@@ -1,63 +1,64 @@
 /* eslint-disable */
-import useLocalStorage from '../src/useLocalStorage';
+import useSessionStorage from '../src/useSessionStorage';
 import { resetStorageState } from '../src/useStorage';
 import 'jest-localstorage-mock';
 import { renderHook, act } from '@testing-library/react-hooks';
 
-describe(useLocalStorage, () => {
+describe(useSessionStorage, () => {
   afterEach(() => {
-    localStorage.clear();
+    sessionStorage.clear();
     jest.clearAllMocks();
     resetStorageState();
   });
 
-  it('retrieves an existing value from localStorage', () => {
-    localStorage.setItem('foo', '"bar"');
-    const { result } = renderHook(() => useLocalStorage('foo'));
+  it('retrieves an existing value from sessionStorage', () => {
+    sessionStorage.setItem('foo', '"bar"');
+    const { result } = renderHook(() => useSessionStorage('foo'));
     const [state] = result.current;
     expect(state).toEqual('bar');
   });
 
-  it('should return initialValue if localStorage empty and set that to localStorage', () => {
-    const { result } = renderHook(() => useLocalStorage('foo', 'bar'));
+  it('should return initialValue if sessionStorage empty and set that to sessionStorage', () => {
+    const { result } = renderHook(() => useSessionStorage('foo', 'bar'));
     const [state] = result.current;
     expect(state).toEqual('bar');
-    expect(localStorage.__STORE__.foo).toEqual('"bar"');
+    expect(sessionStorage.__STORE__.foo).toEqual('"bar"');
   });
 
   it('prefers existing value over initial state', () => {
-    localStorage.setItem('foo', '"bar"');
-    const { result } = renderHook(() => useLocalStorage('foo', 'baz'));
+    sessionStorage.setItem('foo', '"bar"');
+    const { result } = renderHook(() => useSessionStorage('foo', 'baz'));
     const [state] = result.current;
     expect(state).toEqual('bar');
   });
 
-  it('does not clobber existing localStorage with initialState', () => {
-    localStorage.setItem('foo', '"bar"');
-    const { result } = renderHook(() => useLocalStorage('foo', 'buzz'));
+  it('does not clobber existing sessionStorage with initialState', () => {
+    sessionStorage.setItem('foo', '"bar"');
+    const { result } = renderHook(() => useSessionStorage('foo', 'buzz'));
     result.current; // invoke current to make sure things are set
-    expect(localStorage.__STORE__.foo).toEqual('"bar"');
+    expect(sessionStorage.__STORE__.foo).toEqual('"bar"');
   });
 
-  it('correctly updates localStorage', () => {
-    const { result, rerender } = renderHook(() => useLocalStorage('foo', 'bar'));
+  it('correctly updates sessionStorage', () => {
+    const { result, rerender } = renderHook(() => useSessionStorage('foo', 'bar'));
 
     const [, setFoo] = result.current;
     act(() => setFoo('baz'));
     rerender();
 
-    expect(localStorage.__STORE__.foo).toEqual('"baz"');
+    expect(sessionStorage.__STORE__.foo).toEqual('"baz"');
   });
 
-  it('should return undefined if no initialValue provided and localStorage empty', () => {
-    const { result } = renderHook(() => useLocalStorage('some_key'));
+  it('should return undefined if no initialValue provided and sessionStorage empty', () => {
+    const { result } = renderHook(() => useSessionStorage('some_key'));
 
     expect(result.current[0]).toBeUndefined();
   });
 
-  it('returns and allows null setting', () => {
-    localStorage.setItem('foo', 'null');
-    const { result, rerender } = renderHook(() => useLocalStorage('foo'));
+  it('returns and allow setting null', () => {
+    sessionStorage.setItem('foo', 'null');
+    const { result, rerender } = renderHook(() => useSessionStorage('foo'));
+
     const [foo1, setFoo] = result.current;
     act(() => setFoo(null));
     rerender();
@@ -68,12 +69,12 @@ describe(useLocalStorage, () => {
   });
 
   it('sets initialState if initialState is an object', () => {
-    renderHook(() => useLocalStorage('foo', { bar: true }));
-    expect(localStorage.__STORE__.foo).toEqual('{"bar":true}');
+    renderHook(() => useSessionStorage('foo', { bar: true }));
+    expect(sessionStorage.__STORE__.foo).toEqual('{"bar":true}');
   });
 
   it('correctly and promptly returns a new value', () => {
-    const { result, rerender } = renderHook(() => useLocalStorage('foo', 'bar'));
+    const { result, rerender } = renderHook(() => useSessionStorage('foo', 'bar'));
 
     const [, setFoo] = result.current;
     act(() => setFoo('baz'));
@@ -85,9 +86,9 @@ describe(useLocalStorage, () => {
 
   /*
   it('keeps multiple hooks accessing the same key in sync', () => {
-    localStorage.setItem('foo', 'bar');
-    const { result: r1, rerender: rerender1 } = renderHook(() => useLocalStorage('foo'));
-    const { result: r2, rerender: rerender2 } = renderHook(() => useLocalStorage('foo'));
+    sessionStorage.setItem('foo', 'bar');
+    const { result: r1, rerender: rerender1 } = renderHook(() => useSessionStorage('foo'));
+    const { result: r2, rerender: rerender2 } = renderHook(() => useSessionStorage('foo'));
 
     const [, setFoo] = r1.current;
     act(() => setFoo('potato'));
@@ -103,21 +104,21 @@ describe(useLocalStorage, () => {
   });
   */
 
-  it('parses out objects from localStorage', () => {
-    localStorage.setItem('foo', JSON.stringify({ ok: true }));
-    const { result } = renderHook(() => useLocalStorage<{ ok: boolean }>('foo'));
+  it('parses out objects from sessionStorage', () => {
+    sessionStorage.setItem('foo', JSON.stringify({ ok: true }));
+    const { result } = renderHook(() => useSessionStorage<{ ok: boolean }>('foo'));
     const [foo] = result.current;
     expect(foo!.ok).toEqual(true);
   });
 
-  it('safely initializes objects to localStorage', () => {
-    const { result } = renderHook(() => useLocalStorage<{ ok: boolean }>('foo', { ok: true }));
+  it('safely initializes objects to sessionStorage', () => {
+    const { result } = renderHook(() => useSessionStorage<{ ok: boolean }>('foo', { ok: true }));
     const [foo] = result.current;
     expect(foo!.ok).toEqual(true);
   });
 
-  it('safely sets objects to localStorage', () => {
-    const { result, rerender } = renderHook(() => useLocalStorage<{ ok: any }>('foo', { ok: true }));
+  it('safely sets objects to sessionStorage', () => {
+    const { result, rerender } = renderHook(() => useSessionStorage<{ ok: any }>('foo', { ok: true }));
 
     const [, setFoo] = result.current;
     act(() => setFoo({ ok: 'bar' }));
@@ -128,7 +129,7 @@ describe(useLocalStorage, () => {
   });
 
   it('safely returns objects from updates', () => {
-    const { result, rerender } = renderHook(() => useLocalStorage<{ ok: any }>('foo', { ok: true }));
+    const { result, rerender } = renderHook(() => useSessionStorage<{ ok: any }>('foo', { ok: true }));
 
     const [, setFoo] = result.current;
     act(() => setFoo({ ok: 'bar' }));
@@ -139,9 +140,9 @@ describe(useLocalStorage, () => {
     expect(foo!.ok).toEqual('bar');
   });
 
-  it('sets localStorage from the function updater', () => {
+  it('sets sessionStorage from the function updater', () => {
     const { result, rerender } = renderHook(() =>
-      useLocalStorage<{ foo: string; fizz?: string }>('foo', { foo: 'bar' })
+      useSessionStorage<{ foo: string; fizz?: string }>('foo', { foo: 'bar' })
     );
 
     const [, setFoo] = result.current;
@@ -154,7 +155,7 @@ describe(useLocalStorage, () => {
   });
 
   it('rejects nullish or undefined keys', () => {
-    const { result } = renderHook(() => useLocalStorage(null as any));
+    const { result } = renderHook(() => useSessionStorage(null as any));
     try {
       result.current;
       fail('hook should have thrown');
@@ -166,9 +167,9 @@ describe(useLocalStorage, () => {
   /* Enforces proper eslint react-hooks/rules-of-hooks usage */
   describe('eslint react-hooks/rules-of-hooks', () => {
     it('memoizes an object between rerenders', () => {
-      const { result, rerender } = renderHook(() => useLocalStorage('foo', { ok: true }));
+      const { result, rerender } = renderHook(() => useSessionStorage('foo', { ok: true }));
 
-      result.current; // if localStorage isn't set then r1 and r2 will be different
+      result.current; // if sessionStorage isn't set then r1 and r2 will be different
       rerender();
       const [r2] = result.current;
       rerender();
@@ -176,19 +177,19 @@ describe(useLocalStorage, () => {
       expect(r2).toBe(r3);
     });
 
-    it('memoizes an object immediately if localStorage is already set', () => {
-      localStorage.setItem('foo', JSON.stringify({ ok: true }));
-      const { result, rerender } = renderHook(() => useLocalStorage('foo', { ok: true }));
+    it('memoizes an object immediately if sessionStorage is already set', () => {
+      sessionStorage.setItem('foo', JSON.stringify({ ok: true }));
+      const { result, rerender } = renderHook(() => useSessionStorage('foo', { ok: true }));
 
-      const [r1] = result.current; // if localStorage isn't set then r1 and r2 will be different
+      const [r1] = result.current; // if sessionStorage isn't set then r1 and r2 will be different
       rerender();
       const [r2] = result.current;
       expect(r1).toBe(r2);
     });
 
     it('memoizes the setState function', () => {
-      localStorage.setItem('foo', JSON.stringify({ ok: true }));
-      const { result, rerender } = renderHook(() => useLocalStorage('foo', { ok: true }));
+      sessionStorage.setItem('foo', JSON.stringify({ ok: true }));
+      const { result, rerender } = renderHook(() => useSessionStorage('foo', { ok: true }));
       const [, s1] = result.current;
       rerender();
       const [, s2] = result.current;
@@ -197,16 +198,16 @@ describe(useLocalStorage, () => {
   });
 
   describe('Options: raw', () => {
-    it('returns a string when localStorage is a stringified object', () => {
-      localStorage.setItem('foo', JSON.stringify({ fizz: 'buzz' }));
-      const { result } = renderHook(() => useLocalStorage('foo', null, { raw: true }));
+    it('returns a string when sessionStorage is a stringified object', () => {
+      sessionStorage.setItem('foo', JSON.stringify({ fizz: 'buzz' }));
+      const { result } = renderHook(() => useSessionStorage('foo', null, { raw: true }));
       const [foo] = result.current;
       expect(typeof foo).toBe('string');
     });
 
     it('returns a string after an update', () => {
-      localStorage.setItem('foo', JSON.stringify({ fizz: 'buzz' }));
-      const { result, rerender } = renderHook(() => useLocalStorage('foo', null, { raw: true }));
+      sessionStorage.setItem('foo', JSON.stringify({ fizz: 'buzz' }));
+      const { result, rerender } = renderHook(() => useSessionStorage('foo', null, { raw: true }));
 
       const [, setFoo] = result.current;
 
@@ -222,8 +223,8 @@ describe(useLocalStorage, () => {
     });
 
     it('still forces setState to a string', () => {
-      localStorage.setItem('foo', JSON.stringify({ fizz: 'buzz' }));
-      const { result, rerender } = renderHook(() => useLocalStorage('foo', null, { raw: true }));
+      sessionStorage.setItem('foo', JSON.stringify({ fizz: 'buzz' }));
+      const { result, rerender } = renderHook(() => useSessionStorage('foo', null, { raw: true }));
 
       const [, setFoo] = result.current;
 
@@ -237,8 +238,8 @@ describe(useLocalStorage, () => {
   });
 
   it('both components should be updated', () => {
-    const { result: result1 } = renderHook(() => useLocalStorage('foo', 'bar'));
-    const { result: result2 } = renderHook(() => useLocalStorage('foo'));
+    const { result: result1 } = renderHook(() => useSessionStorage('foo', 'bar'));
+    const { result: result2 } = renderHook(() => useSessionStorage('foo'));
     let [ state1, setState1 ] = result1.current;
     let [ state2 ] = result2.current;
     expect(state1).toBe('bar');